Move CLIOptions to its own class
